Removed the getMarshal methods from Marshal in XStrategy.
Refactored the CoreMarshal compiler to avoid the use of getMarhsal in
Madlib.

Created the marshal package as a stomping ground for the new Marshal
classes, these classes include Hook, Launcher, MarhsalRedux,
CoreMarshalRedux, MarshalStrategy, Primer, StrategistRedux, and
DispatchMarhsal.

Commented out a unused class in IndexReduxTestCase for warning
suppression in Strata.

Checked in skeletal CoreSheafTestCase that was lying around in Sheaf.


git-svn-id: http://blogometer.com/svn/open/trunk/strata@583 3df9ba58-24f1-0310-ade5-ba95bd924a04